finder_src/requirements.txt
#Imagens
pillow==9.5.0